Extract annotations for Dex methods
